split hash/encrypt password tests
